Fix bug in IOMMU command buffer handling
While the logic to allow for a small command buffer is correct, the
buffer tail pointer was being used erroneously as an index into the
command_buf array. When command_buf happened to be page aligned due
simply to how the final binary was assembled, it just happened that
the buffer tail pointer would start as zero and would work as an
index. But if the command_buf got aligned differently e.g. only
with 16b alignment, the index would overrun the command_buf
and garbage commands would be programmed into the IOMMU. This would
cause things to hang waiting on command  completion. This
situation was triggered simply by the commit to refactor the code
which changed the layout of the final binary.

The fix is to use a separate index variable initialized to zero
on each call to the iommu_load_device_table(). This index is
passed to send_command() and incremented after it is used.

Fixes: 7aba042f2a6c ("skl: some light refactoring")
Signed-off-by: Ross Philipson <ross.philipson@oracle.com>