Add support for simple payloads
This adds new SKL tag to allow for booting simplest payload. Tag contains
base and size of the payload, its entry point and a single argument. SKL
extends just the payload to PCR17, and argument is not used except for
passing it to the payload using cdecl convention. It is the responsibility
of the payload to measure this argument before using. Payload also has to
set GIF at appropriate time, as well as not use INIT for starting APs, as
usual.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>