CI: Further adjustments to 32 vs 64bit building
It turns out that CodeQL gets confused with an integer 'bits' value in the
matrix, causing it to merge results.  Switch to using strings values instead.

Additionally, M32 is a very unintuitive name.  Change the build system to take
a BITS=32/64 input and parse it appropriately.  Update the build invocations
to match.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>