head.S: make state on entry to the kernel similar to that used by TXT
Intel TXT provides physical address of kernel entry point in %ebx. This
is due to the fact that %ss is undefined there, so the kernel can't use
the call/pop pair to obtain its load address. Even though it is possible
on AMD with this implementation of SKL, keep things consistent and move
the entry point address to %ebx as well.

%ebp points to base of SLB, the kernel can use it to obtain offset to
SLRT, and through it, bootloader context and payload argument saved
within.

Note that this commit (temporarily) breaks booting to non-Linux
payloads.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>